requirements.txt
fastmcp>=0.1.0
msal>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0



